requirements.txt
tensorflow>=2.15.0,<2.18.0
Pillow>=10.0.0
numpy>=1.24.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
gunicorn
gdown>=4.7.1
requests>=2.31.0